Make guide the main page instead of summary
